d08 - HauntedWasteland.kt
I got because I missed that we are starting from AAA and ending with ZZZ. Instead, I was trying to set first element as the start, and the last as the end.

In the second part I did not see that the lcm and gcd arithmetics would be the solution for my problem.